+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2024-11-23
+
+### Added
+- [audio/Raspberry_pi_pico_w_pinout.webp](./audio/Raspberry_pi_pico_w_pinout.webp)
+- [audio/firmware.py](./audio/firmware.py)
+- [audio/src/boot.py](./audio/src/boot.py)
+- [audio/src/requirements.txt](./audio/src/requirements.txt)
+- [audio/src1/boot.py](./audio/src1/boot.py)
+- [audio/src1/requirements.txt](./audio/src1/requirements.txt)
+- [audio/src2/code.py](./audio/src2/code.py)
+- [audio/src2/settings.toml](./audio/src2/settings.toml)
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [0.0.3] - 2024-11-23
 
 ### Added